refactor: Split aggregate ops from implementation
